requirements.txt
aniso8601
certifi
charset-normalizer
click
Flask
Flask-RESTful
idna
itsdangerous
Jinja2
MarkupSafe
marshmallow
packaging
psycopg2-binary
pytz
requests
six
urllib3
Werkzeug
SQLAlchemy
